% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accessors.spdur.R
\name{AIC.spdur}
\alias{AIC.spdur}
\title{AIC method for spdur}
\usage{
\method{AIC}{spdur}(object, ..., k = 2)
}
\arguments{
\item{object}{An object of class \code{spdur}.}

\item{k}{The penalty parameter, by default 2. For \code{\link{BIC.spdur}},
the penalty parameter equals \code{log(N)}.}

\item{\dots}{Optional arguments.}
}
\description{
Computes the Akaike Information Criterion for an \code{spdur} class object.
}
\examples{
data(model.coups)
AIC(model.coups)
}
\seealso{
\code{link{AIC}}, \code{link{BIC.spdur}}
}

