% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/attempt_date.R
\name{attempt_date}
\alias{attempt_date}
\title{Attempt to convert to R date format}
\usage{
attempt_date(date, by)
}
\arguments{
\item{date}{character or numeric}

\item{by}{character, "year", "month", or "day"}
}
\description{
For internal use only
}
\keyword{internal}

