% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.call.R
\name{expand.call}
\alias{expand.call}
\title{Expand call to full names.}
\usage{
expand.call(definition = NULL, call = sys.call(sys.parent(1)),
  expand.dots = TRUE, eval = FALSE)
}
\arguments{
\item{definition}{a function. See \code{\link[base]{match.call}}.}

\item{call}{an unevaluated call to the function specified by definition.
See \code{\link[base]{match.call}}.}

\item{expand.dots}{logical. Should arguments matching ... in the call be 
included or left as a ... argument? See \code{\link[base]{match.call}}.}

\item{eval}{logical, defaults to TRUE. Should function arguments be 
evaluated in the returned call or not?}
}
\value{
An object of class call.
}
\description{
Return a call in which all of the arguments which were supplied
or have presets are specified by their full names and supplied
or default values.
}
\seealso{
\code{\link[base]{match.call}}
}
\author{
fabians
}
\keyword{internal}
