% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.spdur.R
\name{hazard}
\alias{hazard}
\title{Calculate hazard function values}
\usage{
hazard(ti, lambda, cure, alpha, out, dist)
}
\arguments{
\item{ti}{Vector of duration values over which to evaluate the hazard function.}

\item{lambda}{density function parameter}

\item{cure}{density function parameter}

\item{alpha}{density function parameter}

\item{out}{not used}

\item{dist}{"weibull" or "loglog"}
}
\description{
Calculate hazard function values
}
\keyword{internal}
