% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accessors.spdur.R
\name{logLik.spdur}
\alias{logLik.spdur}
\title{Log-Likelihood of an spdur Object}
\usage{
\method{logLik}{spdur}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{spdur}.}

\item{\dots}{not used}
}
\description{
Returns the log-likelihood of a \code{spdur} class object.
}
\examples{
data(model.coups)
logLik(model.coups)
}

