% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sploglog.R
\name{loglog}
\alias{loglog}
\title{Regular Log-logistic regression}
\usage{
loglog(Y, X, inits = NULL, max.iter, silent = TRUE)
}
\arguments{
\item{Y}{Dependent variables.}

\item{X}{Duration equation covariates.}

\item{inits}{Vector of starting values.}

\item{max.iter}{Maximum iterations to try.}

\item{silent}{Supress \code{\link{optim}} trace output?}
}
\description{
Regular Log-logistic regression
}
\keyword{internal}

