% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglog_lnl}
\alias{loglog_lnl}
\title{Regular Log-logistic likelihood}
\usage{
loglog_lnl(theta, y, X)
}
\arguments{
\item{theta}{Current parameter values.}

\item{y}{Dependent variables.}

\item{X}{Duration equation covariates.}
}
\description{
Regular Log-logistic likelihood
}

