% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{model.coups}
\alias{model.coups}
\title{Model of global coups from 1979 to 2010}
\format{An object of class \code{spdur}.}
\source{
For information on the data used in this model, see the data documentation,
\code{\link{coups}}.
}
\usage{
model.coups
}
\description{
This is a model object for a split-duration model of the 
Powell & Thyne coups. It is used in several example code sections to speed
up package testing by eliminating the need to re-estimate a model each time.
}
\examples{
data(model.coups)
str(model.coups)
}
