% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.spdur.R
\name{plot_hazard1}
\alias{plot_hazard1}
\title{Plot hazard rate}
\usage{
plot_hazard1(x, t_lim = NULL, ...)
}
\arguments{
\item{x}{An object of class "\code{spdur}".}

\item{t_lim}{Time period over which to plot the hazard function.}

\item{\dots}{Optional arguments to \code{\link{plot}}.}
}
\description{
Plot hazard rate
}

