% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_hazard.r
\name{plot_hazard2}
\alias{plot_hazard2}
\title{Simulate and plot hazard function}
\usage{
plot_hazard2(x, xvals = NULL, zvals = NULL)
}
\arguments{
\item{x}{An object of class \code{spdur}}

\item{xvals}{A vector of values for the duration equation variables, in the
  same order as the duration equation in \code{x}. Defaults to means.}

\item{zvals}{A vector of values for the risk equation varialbes, in teh same
   order as the risk equation in \code{x}. Defaults to means.}
}
\description{
Given a set of values for the duration and risk equations, plot the shape of
 estimated hazard function in respect to duration. Confidence intervals are
 provided through simulation.
}

