% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sploglog.R
\name{sploglog}
\alias{sploglog}
\title{Split-population Log-logistic regression}
\usage{
sploglog(Y, X, Z, max.iter, silent = FALSE)
}
\arguments{
\item{Y}{Dependent variables.}

\item{X}{Duration equation covariates.}

\item{Z}{Risk equation covariates.}

\item{max.iter}{Maximum iterations to try.}

\item{silent}{Supress \code{\link{optim}} trace output?}
}
\description{
Split-population Log-logistic regression
}
\keyword{internal}

