% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sploglog_lnl}
\alias{sploglog_lnl}
\title{Split-population Log-logistic log likelihood}
\usage{
sploglog_lnl(theta, y, X, Z)
}
\arguments{
\item{theta}{Current parameter values.}

\item{y}{Dependent variables.}

\item{X}{Duration equation covariates.}

\item{Z}{Risk equation covariates.}
}
\description{
Split-population Log-logistic log likelihood
}
\keyword{internal}

