% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spweib_lnl}
\alias{spweib_lnl}
\title{Split-population Weibull log likelihood}
\usage{
spweib_lnl(theta, y, X, Z)
}
\arguments{
\item{theta}{Current parameter values.}

\item{y}{Dependent variables.}

\item{X}{Duration equation covariates.}

\item{Z}{Risk equation covariates.}
}
\description{
Split-population Weibull log likelihood
}
\keyword{internal}
