% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spweibull.R
\name{spweibull}
\alias{spweibull}
\title{Split-population Weibull regression}
\usage{
spweibull(Y, X, Z, max.iter, silent = FALSE)
}
\arguments{
\item{Y}{Dependent variables.}

\item{X}{Duration equation covariates.}

\item{Z}{Risk equation covariates.}

\item{max.iter}{Maximum iterations to try.}

\item{silent}{Supress \code{\link{optim}} trace output?}
}
\description{
Split-population Weibull regression
}
\keyword{internal}

