% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weib_lnl}
\alias{weib_lnl}
\title{Regular weibull log likelihood}
\usage{
weib_lnl(theta, y, X)
}
\arguments{
\item{theta}{Current parameter values.}

\item{y}{Dependent variables.}

\item{X}{Duration equation covariates.}
}
\description{
Regular weibull log likelihood
}
\keyword{internal}
