% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spweibull.R
\name{weibull}
\alias{weibull}
\title{Regular Weibull regression}
\usage{
weibull(Y, X, inits = NULL, max.iter, silent = TRUE)
}
\arguments{
\item{Y}{Dependent variables.}

\item{X}{Duration equation covariates.}

\item{inits}{Vector of starting values.}

\item{max.iter}{Maximum iterations to try.}

\item{silent}{Supress \code{\link{optim}} trace output?}
}
\description{
Regular Weibull regression
}
\keyword{internal}
