% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/attemptDate.R
\name{attemptDate}
\alias{attemptDate}
\title{Attempt to convert to R date format}
\usage{
attemptDate(date, by)
}
\arguments{
\item{date}{character or numeric}

\item{by}{character, "year", "month", or "day"}
}
\description{
For internal use only
}

