% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast.default}
\alias{forecast.default}
\title{Default forecast method}
\usage{
\method{forecast}{default}(object, ...)
}
\arguments{
\item{object}{A model object.}

\item{\dots}{Additional arguments}
}
\description{
The default forecast method, currently undefined.
}

