% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.spdur.R
\name{hazard}
\alias{hazard}
\title{Calculate hazard function values}
\usage{
hazard(x, t = NULL)
}
\arguments{
\item{x}{An object of class "\code{spdur}".}

\item{t}{Vector of duration values over which to evaluate the hazard function.}
}
\description{
Calculate hazard function values
}

