% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.spdur.R
\name{hazard}
\alias{hazard}
\title{Calculate hazard function values}
\usage{
hazard(ti, lambda, cure, alpha, out, dist)
}
\arguments{
\item{ti}{Vector of duration values over which to evaluate the hazard function.}

\item{x}{An object of class "\code{spdur}".}
}
\description{
Calculate hazard function values
}
\keyword{internal}

