% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.spdur.R
\name{plot.spdur}
\alias{plot.spdur}
\title{Plot split-duration model results.}
\usage{
\method{plot}{spdur}(x, type = "sepplot", ...)
}
\arguments{
\item{x}{An object of class "\code{spdur}".}

\item{type}{What kind of plot? "hazard1", "hazard2", "sepplot".}

\item{\dots}{Optional parameters passed to \code{\link{separationplot.spdur}},
  \code{\link{plot_hazard1}} or \code{\link{plot_hazard2}}.}
}
\description{
See \code{\link{separationplot.spdur}} and \code{\link{plot_hazard2}}.
}
\examples{
# get model estimates
data(model.coups)

# plot
p <- plot(model.coups)
}
\seealso{
\code{\link{separationplot.spdur}}, \code{\link{plot_hazard1}},
  \code{\link{plot_hazard2}}
}

