% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.spdur.R
\name{plot_hazard1}
\alias{plot_hazard1}
\title{Plot conditional hazard rate}
\usage{
plot_hazard1(x, ...)
}
\arguments{
\item{x}{class "spdur" object}

\item{...}{passed to \code{plot_hazard}}
}
\value{
NULL, plots.
}
\description{
Plot hazard function without simulated confidence intervals. See
\code{\link{plot_hazard}} instead.
}

