% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.spdur.R
\name{plot_hazard2}
\alias{plot_hazard2}
\title{Simulate and plot hazard function}
\usage{
plot_hazard2(x, ...)
}
\arguments{
\item{x}{class "spdur" object}

\item{...}{passed to \code{plot_hazard}}
}
\value{
NULL, plots.
}
\description{
Plot hazard function with simulated confidence intervals. See
\code{\link{plot_hazard}} instead.
}
